% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsr.R
\name{plot.ips}
\alias{plot.ips}
\title{Visualize the influential scores}
\usage{
\method{plot}{ips}(obj, topn = 5, ylim = NULL)
}
\arguments{
\item{obj}{the objective obtained from rank.compare function}

\item{topn}{the number of influential cases labelled in the plot}

\item{ylim}{y coordinates ranges}
}
\description{
Display the influential score using lollipop plot
}
\examples{
X <- miller05$X
y <- miller05$y

obj <- getdrop1ranks(X, y, topn = 100, fun = "t.test")
rks <-rank.compare(refr = obj$orig, matr = obj$drop1rank, topn = 100, dist = "L2")
par(mfrow = c(1, 2), mar = c(4, 4, 2, 2))
plot_unweighted_ranks(rks, top = TRUE)
plot_kappa_weight(rks$kappa, n = 100, type = "line")
plot(rks, topn = 5, ylim = NULL)
}
\author{
Shuo Wang
}
