% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipsr.R
\name{plot_kappa_weight}
\alias{plot_kappa_weight}
\title{Visualize the weight function}
\usage{
plot_kappa_weight(kappa, n, type = c("line", "points"), ylim = NULL)
}
\arguments{
\item{kappa}{a parameter of the weight function.}

\item{n}{the length list.}

\item{type}{draw line or points.}

\item{ylim}{y coordinates ranges}
}
\description{
Plot the weight function given kappa and the list length (n)
}
\examples{
X <- miller05$X
y <- miller05$y

obj <- getdrop1ranks(X, y, topn = 100, fun = "t.test")
rks <-rank.compare(refr = obj$orig, matr = obj$drop1rank, topn = 100, dist = "L2")
par(mfrow = c(1, 2), mar = c(4, 4, 2, 2))
plot_unweighted_ranks(rks, top = TRUE)
plot_kappa_weight(rks$kappa, n = 100, type = "line")
plot.ips(rks, topn = 5, ylim = NULL) 
}
\author{
Shuo Wang
}
